refactor: editable message
